+++ LibManagemProjBootMVC/src/main/java/ks/m5s/dao/ILibrarianDao.java
@@ -3,6 +3,6 @@ package ks.m5s.dao;
 import org.springframework.data.repository.CrudRepository;
 import ks.m5s.model.Librarian;
 
-public interface ILibrarianDao extends CrudRepository<Librarian, Integer> {
+public interface ILibrarianDao extends CrudRepository<Librarian, Long> {
 
 }
