+++ LibManagemProjBootMVC/src/main/java/ks/m5s/dao/IStudentDao.java
@@ -3,7 +3,7 @@ package ks.m5s.dao;
 import org.springframework.data.repository.CrudRepository;
 import ks.m5s.model.Student;
 
-public interface IStudentDao extends CrudRepository<Student, Integer> {
+public interface IStudentDao extends CrudRepository<Student, Long> {
